+++ internal/util/echo/info.go
