+++ internal/util/execute/execute.go
@@ -6,14 +6,30 @@
 // THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
 // See the Mulan PSL v2 for more details.
 
-package command
+package execute
 
 import (
 	"context"
-	"github.com/Luna-CY/dem/command/utils"
+	"github.com/Luna-CY/dem/internal/util/echo"
+	"os"
+	"os/exec"
 )
 
-// UtilsCommandExecute 执行命令
-func UtilsCommandExecute(ctx context.Context) error {
-	return utils.NewUtilsCommand().ExecuteContext(ctx)
+// RunCommand 执行命令
+func RunCommand(ctx context.Context, working string, command string) error {
+	var commands = []string{"/bin/bash", "-c", command}
+
+	echo.InfoLN(commands)
+	var c = exec.CommandContext(ctx, commands[0], commands[1:]...)
+
+	c.Dir = working
+	for _, env := range os.Environ() {
+		c.Env = append(c.Env, env)
+	}
+
+	c.Stdin = os.Stdin
+	c.Stdout = os.Stdout
+	c.Stderr = os.Stderr
+
+	return c.Run()
 }
